Update docs and internals of  factorizations
